requirements.txt
python-dotenv
fastapi
uvicorn
starlette
pydantic
langchain
langchain-google-genai
langchain-huggingface
langgraph
langgraph-prebuilt
langchain-community
wikipedia
arxiv
tavily-python

